Change to int

+++ internal/entity/book.go
@@ -6,10 +6,10 @@ import (
 
 // Book struct holds entity of book
 type Book struct {
-	ID        int64     `json:"id"`
+	ID        int       `json:"id"`
 	Isbn      string    `json:"isbn"`
 	Title     string    `json:"title"`
-	Price     int32     `json:"price"`
+	Price     int       `json:"price"`
 	CreatedAt time.Time `json:"created_at"`
 	UpdatedAt time.Time `json:"updated_at"`
 }
